refactor: language filtering
Loops through all the subtitles and only add subId, lang, rating, and downloadLink
to dictionary if the lang is the specified language. The dictionary would only contain
information on the subtitle links that matches the specified language, less to loop through.
Previously, it would loop through all subtitles and store all subId, lang, rating,
and downloadLink and would need to loop through the dictionary with all subtitles
and identify only those that matches the specified language.